==> MainCode/0_Codigo_Base/1_Preparacion.ipynb <==
# Preparación de los archivos CSV's

class Analyzer:
    def __init__(self, df, keywords_regex=None):
        self.df = df
        # Si se proporciona una lista de expresiones regulares, conviértela en un diccionario
        if keywords_regex:
            self.keywords_regex = {f'Keyword_{i+1}': regex for i, regex in enumerate(keywords_regex)}
        else:
            # Si no se proporciona, usa un conjunto predeterminado de palabras clave
            self.keywords_regex = {
                'Need': r'need',
                'Keywords': r'keywords'
            }

    def clean_dataframe(self):
        # Conservar solo la columna 'Abstract'
        self.df = self.df[['Abstract']]
        # Convertir a minúsculas
        self.df['Abstract'] = self.df['Abstract'].apply(lambda x: x.lower())

    def find_keywords(self):
        # Encontrar las coincidencias para cada palabra clave
        keyword_matches = {}
        for keyword, regex in self.keywords_regex.items():
            keyword_matches[keyword] = self.df['Abstract'].apply(lambda x: bool(re.search(regex, x)))
        
        # Filtrar las filas que contienen todas las palabras clave
        valid_rows = self.df
        for match in keyword_matches.values():
            valid_rows = valid_rows[match]
        
        return valid_rows
